Tipo de dato cambiando para el puerto

+++ src/settings/settings.go
@@ -14,7 +14,7 @@ var DB *Database
 
 type Database struct {
 	Host     *string `yaml:"host"`
-	Port     *int    `yaml:"port"`
+	Port     *uint16 `yaml:"port"`
 	User     *string `yaml:"user"`
 	Password *string `yaml:"password"`
 	Name     *string `yaml:"name"`
